requirements.txt
datasets==3.6.0
fastapi==0.115.12
gradio==5.25.2
huggingface_hub==0.31.2
matplotlib==3.10.1
medmnist==3.0.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
Pillow==11.1.0
prometheus-client==0.21.1
python-multipart==0.0.20
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
torch==2.6.0
torchvision==0.21.0
uvicorn[standard]==0.34.1
